refactor: use persistent streams in iroh connections
Previously a uni-directional stream was created and destroyed for each message. That required a fixed delay to be awaited to make a preflight frame possible.
Now the preflight is sent as a first frame on a newly opened connection and stream and the stream is kept alive until an error occurs or the connection is closed.